# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Holivita AI API",
        "description": "API для онбординг-чата с AI-ассистентом, суммаризации и голосового ввода (OpenAI Whisper). Авторизация отключена для MVP — user_id передаётся вручную. Модули: Onboarding (синхронный и асинхронный чат), Summaries (суммаризации сессий), Voice (транскрипция аудио). Формат ответов: успех — {success: true, data: {...}}, ошибка — {success: false, error: \"...\"}.",
        "contact": {
            "name": "Holivita AI Team",
            "email": "support@holivita.ai"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API сервер"
        }
    ],
    "paths": {
...
        "/summaries": {
            "get": {
                "tags": [
                    "Summaries"
                ],
                "summary": "Список суммаризаций завершённых сессий",
                "description": "Возвращает пагинированный список суммаризаций завершённых сессий онбординга. Можно фильтровать по user_id. Суммаризация содержит структурированные данные о пользователе, собранные в ходе онбординга. Сортировка по дате завершения (новые первые).",
                "operationId": "summariesIndex",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "Фильтр по ID пользователя (если не указан — возвращаются все суммаризации)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "example": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Количество элементов на страницу (от 1 до 100, по умолчанию 15)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15,
                            "maximum": 100,
                            "minimum": 1,
                            "example": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список суммаризаций с пагинацией",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "items": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "string",
                                                                "format": "uuid",
                                                                "example": "550e8400-e29b-41d4-a716-446655440000"
                                                            },
                                                            "user_id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "summary": {
                                                                "description": "Структурированная суммаризация данных пользователя",
                                                                "type": "object"
                                                            },
                                                            "completed_at": {
                                                                "type": "string",
                                                                "format": "date-time",
                                                                "example": "2026-02-10T12:00:00.000000Z"
                                                            },
                                                            "created_at": {
                                                                "type": "string",
                                                                "format": "date-time",
                                                                "example": "2026-02-10T11:30:00.000000Z"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "pagination": {
                                                    "properties": {
                                                        "current_page": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "last_page": {
                                                            "type": "integer",
                                                            "example": 3
                                                        },
                                                        "per_page": {
                                                            "type": "integer",
                                                            "example": 15
                                                        },
                                                        "total": {
                                                            "type": "integer",
                                                            "example": 42
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/summaries/{sessionId}": {
            "get": {
                "tags": [
                    "Summaries"
                ],
                "summary": "Получение суммаризации по ID сессии",
                "description": "Возвращает суммаризацию конкретной завершённой сессии онбординга. Суммаризация содержит структурированные данные (JSON), собранные AI-ассистентом в ходе диалога с пользователем.",
                "operationId": "summariesShow",
                "parameters": [
                    {
                        "name": "sessionId",
                        "in": "path",
                        "description": "UUID сессии онбординга",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "550e8400-e29b-41d4-a716-446655440000"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Суммаризация найдена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "format": "uuid",
                                                    "example": "550e8400-e29b-41d4-a716-446655440000"
                                                },
                                                "user_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "summary": {
                                                    "description": "Структурированная суммаризация данных пользователя",
                                                    "type": "object"
                                                },
                                                "completed_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2026-02-10T12:00:00.000000Z"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2026-02-10T11:30:00.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Суммаризация не найдена (сессия не существует или не завершена)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Суммаризация не найдена."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}